feat: scaffold pipeline bugfixes + regenerate engine and integrations
Five bugs fixed in the scaffold pipeline:
- Interface extends extraction (extends_interfaces node lookup + tc variable)
- Modifier/annotation extraction (is_static, is_final, is_abstract, permits)
- Builder semantics (static→ClassVar, abstract→ABC+@abstractmethod, Protocol always on interfaces)
- Import collection (recursive full-tree walk for all 9 import names)
- Alias resolution (resolve_with_alias + _apply_aliases rewrites type references)

Regenerated 2,921 scaffolds (1,276 engine + 1,645 integrations) with zero errors.
29 regression tests covering all five bugfix categories.

+++ integrations/apify/actor/actor_run_options.py
@@ -1,7 +1,9 @@
 from __future__ import annotations
 
-from dataclasses import dataclass, field
-from typing import Any
+# Source: E:\KESTRA-IO\plugins\plugin-apify\src\main\java\io\kestra\plugin\apify\actor\ActorRunOptions.java
+
+from dataclasses import dataclass
+from typing import Any, Optional
 
 
 @dataclass(slots=True, kw_only=True)
@@ -12,5 +14,5 @@ class ActorRunOptions:
     disk_mbytes: float | None = None
     max_total_charge_usd: float | None = None
 
-    def get_max_total_charge_usd(self) -> Optional[Double]:
+    def get_max_total_charge_usd(self) -> Optional[float]:
         raise NotImplementedError  # TODO: translate from Java
