+++ integrations/apify/actor/pricinginfo/flat_price_per_month_actor_pricing_info.py
@@ -1,7 +1,9 @@
 from __future__ import annotations
 
-from dataclasses import dataclass, field
-from typing import Any
+# Source: E:\KESTRA-IO\plugins\plugin-apify\src\main\java\io\kestra\plugin\apify\actor\pricinginfo\FlatPricePerMonthActorPricingInfo.java
+
+from dataclasses import dataclass
+from typing import Any, Optional
 
 from integrations.apify.actor.pricinginfo.common_actor_pricing_info import CommonActorPricingInfo
 
@@ -11,5 +13,5 @@ class FlatPricePerMonthActorPricingInfo(CommonActorPricingInfo):
     trial_minutes: float | None = None
     price_per_unit_usd: float | None = None
 
-    def get_trial_minutes(self) -> Optional[Double]:
+    def get_trial_minutes(self) -> Optional[float]:
         raise NotImplementedError  # TODO: translate from Java
