+++ integrations/apify/actor/pricinginfo/price_per_event_actor_pricing_info.py
@@ -1,7 +1,9 @@
 from __future__ import annotations
 
-from dataclasses import dataclass, field
-from typing import Any
+# Source: E:\KESTRA-IO\plugins\plugin-apify\src\main\java\io\kestra\plugin\apify\actor\pricinginfo\PricePerEventActorPricingInfo.java
+
+from dataclasses import dataclass
+from typing import Any, Optional
 
 from integrations.apify.actor.pricinginfo.common_actor_pricing_info import CommonActorPricingInfo
 from integrations.apify.actor.pricinginfo.pricing_per_event import PricingPerEvent
@@ -12,5 +14,5 @@ class PricePerEventActorPricingInfo(CommonActorPricingInfo):
     pricing_per_event: PricingPerEvent | None = None
     minimal_max_total_charge_usd: float | None = None
 
-    def get_minimal_max_total_charge_usd(self) -> Optional[Double]:
+    def get_minimal_max_total_charge_usd(self) -> Optional[float]:
         raise NotImplementedError  # TODO: translate from Java
