+++ integrations/mqtt/services/message.py
@@ -1,6 +1,8 @@
 from __future__ import annotations
 
-from dataclasses import dataclass, field
+# Source: E:\KESTRA-IO\plugins\plugin-mqtt\src\main\java\io\kestra\plugin\mqtt\services\Message.java
+
+from dataclasses import dataclass
 from typing import Any
 
 
@@ -9,6 +11,6 @@ class Message:
     id: int | None = None
     topic: str | None = None
     qos: int | None = None
-    properties: list[Byte] | None = None
+    properties: list[int] | None = None
     payload: Any | None = None
     retain: bool | None = None
